+++ src/main.py
@@ -17,6 +17,7 @@ from .routers import (
     game_types,
     injuries,
     login,
+    past_bets,
     predictions,
     reddit_comments,
     schedule,
@@ -43,6 +44,7 @@ app.include_router(feedback.router)
 app.include_router(game_types.router)
 app.include_router(injuries.router)
 app.include_router(login.router)
+app.include_router(past_bets.router)
 app.include_router(predictions.router)
 app.include_router(reddit_comments.router)
 app.include_router(schedule.router)
