+++ templates/bets.html
@@ -33,8 +33,7 @@
 
 <center>
    <h4 class="text-center my-3 pb-3">{{ username }}</h4>
-   <h4 class="text-center my-3 pb-3">NBA Betting</h4>
-   <a href="/past_bets">Past Bets Page</a>
+   <h1 class="text-center my-3 pb-3">NBA User Prediction Bets Page</h4>
 </center>
 <form name="selected_team_winner" action="/bets" method="post" enctype="multipart/form-data">
    <table class="table mb-4">
@@ -87,6 +86,9 @@
          {% endif %}
       </div>
 </form>
+<br>
+<br>
+<a href="/past_bets">Past Bets Page</a>
 </center>
 <!-- <center> <a class="btn btn-success text-align-center" href="/bets">Save</a> </center> -->
