+++ templates/past_bets.html
@@ -2,6 +2,33 @@
 
 {% block content %}
 <center>
+   <br>
+   <h4 class="text-center my-3 pb-3">{{ username }}</h4>
+   <h1 class="text-center my-3 pb-3">NBA User Prediction Past Bets Page</h4>
+   <br>
+
+   <table class="table">
+      <thead>
+         <tr>
+            <th scope="col">Total Correct Predictions</th>
+            <th scope="col">Total Predictions</th>
+            <th scope="col">% Correct</th>
+         </tr>
+      </thead>
+      <tbody>
+         <tr>
+            <th>  {{ past_predictions_success_count }}
+               <td> {{ past_predictions_total_count }} </td>
+               <td> {{ (past_predictions_pct_count * 100)|round(3) }}% </td>
+            </th>
+         </tr>
+      </tbody>
+      </table>
+
+      <br>
+      <br>
+      <br>
+   
    <table class="table">
       <thead>
          <tr>
@@ -36,28 +63,7 @@
    </table>
 
    <br>
-   <br>
-   <br>
-
-   <table class="table">
-   <thead>
-      <tr>
-         <th scope="col">Prediction Type</th>
-         <th scope="col">Count</th>
-         <th scope="col">% Count</th>
-      </tr>
-   </thead>
-   <tbody>
-      <tr>
-         <th> hi 1
-            <td> hi 2 </td>
-            <td> hi 3 </td>
-         </th>
-      </tr>
-   </tbody>
-   </table>
-   
-   <br>
    <a href="/bets">Tonight's Bets Page</a>
+
 </center>
 {% endblock content %}
